Allows the player to change game difficulty at any time.
The KBounceGameWidget::setGameDifficulty() method has been created based
on the old version of the levelChanged() slot and is responsible for
increasing or decreasing balls' and walls' velocity.

In the other hand, levelChanged's new version reacts to the user input
making calls to setGameDifficulty() starting a new game when required.

The point is to make methods follow the single responsibility model,
making things easy to understand and maintain.

BUG: 308889
CCBUG: 308889
REVIEW: 107084