Scene and board positioning improved.
The scene now has the size of the whole game widget, making easier the
process of positioning the board during each resize event.

The background is now rendered in the scene, not in the game widget as
it was being done before. I don't think it's the widget's business to
care about rendering anything, it's the scene work to render objects.

A different commit wasn't made to deal exclusively with the background
rendering part because without improving the scene positioning, the
background wouldn't be rendered as expected, clipping part of its
content.