Don't add empty scores to the highscores table.
As said in the descprition, this change avoids the inclusion of empty
scores, i.e.: when it's equal to zero, in the high scores list.

It doesn't make any sense to threat zero as a valid score, let alone
write it in the high scores table (which is meant for exposing the
player's best matches).

BUG: 317907
FIXED-IN: 4.10
REVIEW: 109999