Improvements to the saveGame() method.
Now the method verifies whether the player is not trying to overwrite a
previously saved game. If that's the case, then a yes or no dialog asks
for user input.

The code also received some cleanup regarding variable names and the
strings spacing.

And last, as that message box appearing all the time would bother even
the most patient person I added a "Don't show this a again." checkbox to
it.

CCMAIL: kamikazeanuj@gmail.com