Don't add new highscore after saving a game.
The highscores are calculated only after the game is over.

In order to achieve this a new method had to be included in
KBounceGameWidget to redraw the scene after the game is saved. This way
we now have a different way to close the game and perform the necessary
adjustments when the game is closed after saving.

CCMAIL: kamikazeanuj@gmail.com