Allow player to save the game at any time.
Previously the game had to be closed after saving the current game. That
was intended to avoid cheating, but at the current state of KBounce it
makes no sense at all.

As the player is not allowed to saved a game when it is over, the
cheating process we've foreseen is no longer possible.